refactor(ui): extract RadioCard component
