+++ renderer/components/UI/index.js
@@ -45,6 +45,7 @@ export Panel, { PanelHeader, PanelBody, PanelFooter } from './Panel'
 export ProgressBar from './ProgressBar'
 export QRCode from './QRCode'
 export Radio from './Radio'
+export RadioCard from './RadioCard'
 export RadioGroup from './RadioGroup'
 export Range from './Range'
 export RowLabel from './RowLabel'
